Customizable Sender
In response to the need for the ability to use custom response send and log logic, I've
defined a SendHandler function that allows people to swap in custom sender
functions which JSHAPI will leverage.

A default sender is still provided and used by default. Also added an
enhancement to go-json-spec-handler to further simplify status checking for
sendable ErrorTypes.